Improve AI order
AI order now completes a row or column which contains three pieces,
and three unoccupied spaces (assuming there are two free spaces on
one side, and only one on the other). If that was confusing, it
looks like this:
(free)(free)(red)(red)(red)(free)
and turns it into this
(free)(red) (red)(red)(red)(free)
Note that by doing this, AI order will win on it's next turn no
matter where the player places their next piece.
The AI is able to do this with both colors, but can't do it
diagonally yet.